+++ src/compiler/llvm_codegen.c
@@ -1004,6 +1004,7 @@ static void llvm_emit_type_decls(GenContext *context, Decl *decl)
 		case DECL_UNION:
 		case DECL_ENUM:
 		case DECL_BITSTRUCT:
+		case DECL_CONST_ENUM:
 			llvm_get_typeid(context, decl->type);
 			break;
 	}
@@ -1360,6 +1361,7 @@ LLVMValueRef llvm_get_ref(GenContext *c, Decl *decl)
 		case DECL_CT_ASSERT:
 		case DECL_DISTINCT:
 		case DECL_ENUM:
+		case DECL_CONST_ENUM:
 		case DECL_ENUM_CONSTANT:
 		case DECL_IMPORT:
 		case DECL_LABEL: