+++ src/compiler/sema_name_resolution.c
@@ -951,6 +951,7 @@ bool sema_resolve_type_decl(SemaContext *context, Type *type)
 			if (!type->function.prototype && type->function.decl->decl_kind == DECL_FNTYPE) return sema_analyse_decl(context, type->function.decl);
 			return true;
 		case TYPE_ENUM:
+		case TYPE_CONST_ENUM:
 		case TYPE_STRUCT:
 		case TYPE_UNION:
 		case TYPE_BITSTRUCT: